+++ src/main/java/guru/springframework/recipe/project/recipeproject/exceptions/ControllerExceptionHandler.java
@@ -1,0 +1,29 @@
+package guru.springframework.recipe.project.recipeproject.exceptions;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.servlet.ModelAndView;
+
+@Slf4j
+@ControllerAdvice
+public class ControllerExceptionHandler {
+
+    //Add response status here other 200 will be returned not 404
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(NumberFormatException.class)
+    public ModelAndView handleNumberFormatException(Exception exception) {
+        log.error("Handling Number Format exceptionError");
+        log.error(exception.getMessage());
+
+        ModelAndView modelAndView = new ModelAndView();
+
+        modelAndView.setViewName("400error");
+        modelAndView.addObject("exception", exception);
+
+        return modelAndView;
+
+    }
+}
